Print per baudrate channel power of elements
Per channel power prints were previously only showing the noiseless
reference channel power and only an average power.
With this change, we add a new information on the print:
the average total power (signal + noise + non-linear noise).
If there are several baudrate propagating, the average per baudrate
is displayed.
For this purpose, baudrate and total power are recorded in each element
upon propagation.

Note that the difference between this total power and the existing
channel power represents the added noise for the considered OMS.
Indeed ROADMs equalize per channel total power, so that power displayed
in 'total pch (dBm)' may contain some noise contribution accumulated with
previous propagation.
Because 'pch out (dBm)' is for the noiseless reference, it is exactly
set to the target power and 'total pch (dBm)' is always matching
'pch out (dBm)' in ROADM prints.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: Idc8d949386a4826123c21a60008b50c51ac182e2