CI: docker: releases should update the `latest` tag
We've been accidentally not updating the `latest` tag on Docker hub
after switching from Travis CI to GitHub actions. Traditionally, we've
assumed the `latest` points to the "latest release", so this patch makes
it simple and will call any pushed "version tag" as the `latest`. This
will become a problem if/when we start maintaining multiple releases,
but I think it's a safe approach for now.

As before, there's also the `master` tag which always points to the,
well, master branch of the repo.

Change-Id: I14c08b12010986d4327bd9d685619a98cfec370f