Check regen_list consistency and remove regen from list for computation
raise errors in case of mismatch between regen_list and regenerator nodes
in the topology.
Remove regenerators from the constraint node list and replace them with
their attached ROADM in order to use the path computation tools as they are.
Regenerators would lead to loops in the path and current functions would
not work. An explicit path with regen rxa-roadmb-regenb-trxc means that
there is a regeneration in roadmb, and that any path going through this
regen would form a loop in roadmb. This would prevent to use networkx tools
such as all_simple_paths, shortest_simple_paths.
Instead using regen rxa-roadmb-trxc, allows to find the path
trxa-roadma-fiber1-roadmb-fiber2-roadmc-trx
and then we just need to restore regenb in the path for propagation, and
mode selection.

This change introduce both functions.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I54a09118ac2f556d739177cf63dc07ff43b30f3f