Calculate CD and PMD penalty
The penalty is calculated and presented separately from the GSNR.

It is also taken into account when optimizing trx mode and verifying
path feasibility in path_requests_run processing.

Specification of CD and PMD penalty is introduced in eqpt_config files,
both as part of trx modes and default SI. Penalty is specified in terms
of three parameters: threshold, limit and osnr_penalty:

- Below threshold, penalty is 0.
- Between threshold and limit, penalty is linearly interpolated
  between 0 and osnr_penalty.
- Above limit, transmission is considered not feasible.

This is in line with how CD and PMD penalty is specified in OpenROADM
and compatible with specifications from most other organizations and
vendors.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I3745eba48ca60c0e4c904839a99b59104eae9216