C+L: introduce multi band amps
Introduce a new multi-band type that contains a list of amps
- reads multiple amps out of the element config.
- deduce frequency band from the amp in the list (not from the list
of type varieties, because this one can hold LOW and HIGH gain type
varieties in case of switched gain)
- autodesign not done for multiband case yet: settings of multiband amps
computed without global view on tilt/SRS/... If element is not explicitely
set as multi-band (thanks to type-variety), multiband amps are not selected
during autodesign, and if multi-band amps have no type_variety, design fails (multiband
auto design not supported yet).
- filter out spectrum according to amps band.
- if the si band (used for design) is bigger than the amplifiers bands, then warns
the user that there is an inconsistent definition.
- don't add a preamp or booster if a multiband amp is already defined.
- remove condition on pspani for the addition of spectrum: indeed
p_span_i could be different for two different bands
- Update amp spectrum determination with multiband case: Multiband amps moves from
f_min-fmax to a list of bands, so the bitmap should be updated accordingly.
so instead of [00011...1110000] bitmap may be [00011...111000000011...1110000]
- use amplifiers band to propagate initial spectrum. These is applied to all amps.
Use amplifier bands to filter si defined by user: filter out based on amps on the path
- propagated nb of channels then depends on the path. This changes invocation results
layout

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I5dd13986efad53300d357cc66ccdda50b289b60b

C+L: use amplifier band to propagate initial spectrum

Use amplifier bands to create si: filter out based on amps on the path
Remove the print of channel number from equipment, since the channel number
may now depend on the path's amplifiers, and reproduce the print in
the function

This changes invocation results layout

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I93db1b645430e922aed71c092a833b9112e1455c