C+L: Reads amp freq band instead of defaults band
explanation for the default Amp f_min:
Objective is to use amplifier band to bound the possible frequencies
to be propagated. Since the current default f_min of Amp in json_io.py is
higher than the SI one, this would result in a different nb of channels
than currently used in tests, and a change in all tests. In order to
avoid this, I preferred to change this value and have consistency
between SI f_min and Amp f_min

Add a set of functions to use amp bands as the useable spectrum on each OMS

change the OMS generation to make use of the amp band

Add filtering functions (demux and mux) to filter out spectrum which is
not in the element band

Filter out spectrum when propagating in Amplifier

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I225b2b2dc0e1f1992c0460f6e08fa9c9bc641edf