Fix error message when spectrum exceeds interval defining fiber loss
The loss_coef_func method of the Fiber class assumes that input
frequency is an array. If frequency is a scalar (such as when
loss_coef_func is called from the loss method), and SpectrumError is
raised because frequency is outside the interval used to define fiber
loss, this results in a crash because the error message tries to index
into frequency. This patch enforced that frequency is an array.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I1dcf53b092d5236d5def6c574ffd44b73f54845f