Add to transmission-example 2 optional arguments, path and route_id:
the --path argument shall contain the path elements in quotes, and separated by '|' or ',',
the --route_id shall contain the specific route_id defined mandatorily in an associated Service-file, later is invoqued with the --service argument.
For both optional arguments, if the source and/or destination is(are) omitted, the missing source and/or sink will be inferred from the provided path elements, or from the related route if applicable.
The route_id argument is given priority over the path argument.
In case a Service-file without a specific route_id is passed, then the default route_id will be the first one found in the Service-file.

Signed-off-by: Florian FRANK <florian1.frank@orange.com>
Change-Id: Id23489e52d90052b5b142e3951956499ce97cacc