Allow explicit path or service for transmission-example
In meshed topologies, this feature eases the usage of transmission-example,
as it allows users to specify the path, or the service in the shell.

Therefore 2 optional arguments are added to transmission-example:
- the --path argument shall contain the path elements in quotes, and separated by |
- the --route_id shall contain the specific route_id defined mandatorily in an associated Service-file,
later is invoqued with the --service argument.
For both, if the source and/or destination is(are) omitted, the missing source and/or sink
is inferred from the provided path elements, or from the related route if applicable.
The route_id argument is given priority over the path argument.

Signed-off-by: Florian FRANK <florian1.frank@orange.com>
Change-Id: Id23489e52d90052b5b142e3951956499ce97cacc