Allow explicit path or service for transmission-example
gnpy-transmission-example only uses shortest path. this patch allows to specify a path
different from the shortest path.

- defined as list of ROADM nodes
- or using the same service file used by gnpy-path-request

Therefore 2 optional arguments are added to transmission-example:

- the --path argument shall contain the path elements separated by |
- or the --service argument, containing the service file path, associated
  with the --route_id argument with the route_id (request_id) defined
  mandatorily in the service file.

source and/or destination are inferred from the provided path elements
 or from the related route if applicable.

The route_id argument is given priority over the path argument.

Compared to gnpy-path-request script, this feature enables to have a deep insight
on what is happening in each elements for a given service request_id and also to
enable a mixed rate scenario on this path.

Co-authored-by: JennyLescop <jenny.lescop@orange.com>

Signed-off-by: Florian FRANK <florian1.frank@orange.com>
Change-Id: Id23489e52d90052b5b142e3951956499ce97cacc