Add tx_osnr in spectral information
If tx_osnr is defined via spectrum then use it to define a tx_osnr per
carrier in si. If spectrum is not defined then use the , else use the
tx_osnr of request to set tx_osnr of si.

Then, the propagate function for requests is changed to update OSNR with
tx_OSNR per carrier defined in si.

This change enables to use a different tx_osnr per carrier.

TODO: The tx_osnr defined in spectrum is not yet taken into account for
the propagate_and_optimize function, because the loop that optimizes
the choice for the mode only loops on baudrate.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I0fcdf559d4f1f8f0047faa257076084ec7adcc77