Refactor: remove variety_list variable which is never used
Probably used from a previous intention to define an set of possible EDFA
in each element instance and leave GNPy choose the best among this list.
However the excel/json part of it ha never been implemented.

There is already a way to restrict amplifier choice thanks to
the "allowed_for_design" flag in each amp. This on applies for all
amp instances though.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I65c05495a15de1e1a80749e035af5da768fd4b3c