Add a frequency-dependent fiber model declared as a list in the equipment-file.
This fiber model is instanciated in the topology-file, and can be customized as follows:
the legacy 'loss_coef'-scalar is used to offset the instantiated fiber model
for the given 'loss_coef_offset_frequency' such that at the 'loss_coef_offset_frequency'
the model values the 'loss_coef'-scalar.
In case a list-model is declared as well, the legacy dictionary-based model has priority.

Signed-off-by: Florian FRANK <florian1.frank@orange.com>
Change-Id: Id4267c2bc78d823031e7a69b11115eb93df39183